requirements.txt
asgiref
certifi
cloudinary
dj-database-url
Django==4.2.11; python_version >= "3.8" and python_version < "3.10"
django-crispy-forms
django-heroku
djangorestframework
gunicorn==22.0.0
pytz
six
sqlparse
tzdata
urllib3
whitenoise
wincertstore
reportlab
psycopg2
markdown
django-filter
xhtml2pdf
coverage
google-generativeai
google-cloud-speech